feat: remove chain, use options instead
